# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "https://gitlab.paradeum.com/pld/cloud-batch",
        "contact": {}
    },
    "paths": {
        "/auth/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "post Auth",
                "parameters": [
                    {
                        "description": "auth",
                        "name": "auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Auth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseString"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseString"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseString"
                        }
                    }
                }
            }
        },
        "/auth/passwd": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Update password",
                "parameters": [
                    {
                        "description": "updateAuth",
                        "name": "updateAuth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateAuth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseString"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseString"
                        }
                    },
                    "401": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseString"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseString"
                        }
                    }
                }
            }
        },
        "/batch/servers": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Batch Create Servers",
                "parameters": [
                    {
                        "description": "batchCreateServers",
                        "name": "batchCreateServers",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BatchCreateServersForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseString"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseString"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseString"
                        }
                    }
                }
            }
        },
        "/servers": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Batch Get Servers",
                "parameters": [
                    {
                        "type": "string",
                        "default": "bfs",
                        "description": "project",
                        "name": "project",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "aliyun",
                        "description": "provider",
                        "name": "provider",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "batch_number",
                        "name": "batch_number",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseString"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseString"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseString"
                        }
                    }
                }
            }
        },
        "/version": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Version"
                ],
                "summary": "Get Version",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseString"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseString"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.ResponseString": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                },
                "errDetail": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                }
            }
        },
        "models.Auth": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5
                },
                "username": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5
                }
            }
        },
        "models.BatchCreateServersForm": {
            "type": "object",
            "required": [
                "count",
                "cpu_core_count",
                "memory_size_mb",
                "project",
                "provider",
                "region_scope"
            ],
            "properties": {
                "count": {
                    "description": "创建数量",
                    "type": "integer",
                    "default": 1,
                    "enum": [
                        1,
                        2,
                        5,
                        10,
                        20,
                        50,
                        100
                    ]
                },
                "cpu_core_count": {
                    "description": "cpu 核数",
                    "type": "integer",
                    "default": 4,
                    "enum": [
                        1,
                        2,
                        4,
                        8,
                        16
                    ]
                },
                "memory_size_mb": {
                    "description": "内存（单位 M)",
                    "type": "integer",
                    "default": 8192,
                    "enum": [
                        512,
                        1024,
                        2048,
                        4096,
                        8192,
                        16384
                    ]
                },
                "project": {
                    "description": "项目标签",
                    "type": "string",
                    "default": "bfs",
                    "enum": [
                        "bfs",
                        "dfs"
                    ]
                },
                "provider": {
                    "description": "云提供者",
                    "type": "string",
                    "default": "aliyun",
                    "enum": [
                        "aliyun"
                    ]
                },
                "region_scope": {
                    "description": "区域前缀: all 全部, cnml 中国大陆, cn 中国， ap 亚太， us 美洲， eu 欧洲, me 中东",
                    "type": "string",
                    "default": "cnml",
                    "enum": [
                        "all",
                        "cnml",
                        "cn",
                        "us",
                        "ap",
                        "me",
                        "eu"
                    ]
                }
            }
        },
        "models.UpdateAuth": {
            "type": "object",
            "required": [
                "oldPassword",
                "password",
                "username"
            ],
            "properties": {
                "oldPassword": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5
                },
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 8
                },
                "username": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}